Move parseEnumeratedAttribute to Element class
