+++ packages/alfa-dom/test/node/attribute.spec.ts
@@ -1,13 +1,9 @@
-import { Map } from "@siteimprove/alfa-map";
-import { None, Some } from "@siteimprove/alfa-option";
+import { None } from "@siteimprove/alfa-option";
 import { Parser } from "@siteimprove/alfa-parser";
 import { Err, Ok } from "@siteimprove/alfa-result";
 import { test } from "@siteimprove/alfa-test";
 
-import { Attribute, Element } from "../../src";
-import { EnumeratedValueError } from "../../src/common/microsyntaxes";
-
-const { parseEnumeratedAttribute } = Element;
+import { Attribute } from "../../src";
 
 const makeAttribute = (str: string) => Attribute.of(None, None, "dummy", str);
 
@@ -18,39 +14,6 @@ const parser: Parser<string, number, string> = (str) =>
 test("parse attributes", (t) => {
   t.deepEqual((makeAttribute("12345").parse(parser)), Ok.of(5));
   t.deepEqual((makeAttribute("").parse(parser)), tooShort);
-});
-
-test("parse enumerated attribute according to specs", (t) => {
-  function enumerated(value: string): Element {
-    return Element.of(None, None, "dummy", (elt) => [
-      Attribute.of(None, None, "enumerated", value),
-    ]);
-  }
-  const noenum = Element.of(None, None, "dummy");
-  const noDefault = Map.from([
-    ["foo", 1],
-    ["bar", 2],
-  ]);
-  const withDefault = Map.from([
-    ["foo", 1],
-    ["bar", 2],
-    [EnumeratedValueError.Missing, 0],
-    [EnumeratedValueError.Invalid, 42],
-  ]);
-
-  const parserNoDefault = parseEnumeratedAttribute("enumerated", noDefault);
-  const parserWithDefault = parseEnumeratedAttribute("enumerated", withDefault);
-
-  t.deepEqual(parserNoDefault(enumerated("Foo")), Some.of(1));
-  t.deepEqual(parserNoDefault(enumerated("this is totally invalid")), None);
-  t.deepEqual(parserNoDefault(noenum), None);
-
-  t.deepEqual(parserWithDefault(enumerated("bAR")), Some.of(2));
-  t.deepEqual(
-    parserWithDefault(enumerated("this is totally invalid")),
-    Some.of(42)
-  );
-  t.deepEqual(parserWithDefault(noenum), Some.of(0));
 });
 
 test("parse space separated token list", (t) => {
