+++ packages/alfa-dom/tsconfig.json
@@ -7,7 +7,6 @@
     "src/jsx.ts",
     "src/namespace.ts",
     "src/common/index.ts",
-    "src/common/microsyntaxes.ts",
     "src/common/predicate.ts",
     "src/node.ts",
     "src/node/attribute.ts",
@@ -35,8 +34,8 @@
     "src/style/rule/style.ts",
     "src/style/rule/supports.ts",
     "src/style/sheet.ts",
-    "test/common/microsyntaxes.spec.ts",
-    "test/node/attribute.spec.ts"
+    "test/node/attribute.spec.ts",
+    "test/node/element.spec.ts"
   ],
   "references": [
     {
